Report config file parsing diagnostics correctly Fixes #36515
