+++ tests/baselines/reference/tsbuild/configFileErrors/no-change-run/reports-syntax-errors-in-config-file.js
@@ -1,5 +1,6 @@
 //// [/lib/no-change-runOutput.txt]
 /lib/tsc --b /src/tsconfig.json
-exitCode:: ExitStatus.Success
+src/tsconfig.json(7,9): error TS1005: ',' expected.
+exitCode:: ExitStatus.DiagnosticsPresent_OutputsSkipped
